Robust input coercion
Collapses bools and integers, magically cuts reals into 10 or less
equally-spaced bins, checks for NAs (which is costly).

The difference between treating reals and integers can be confusing,
but I doubt there is a "best" solution.

Quite likely a "high-perf" switch is needed to disable checks for
a performance gain; though, it would mean a possible DoS as NA in any
factor would trigger segfault, as before this commit.